refactor: turn spotless action into an option
instead of subcommand apply/check we now have --mode=APPLY or --mode=CHECK
