+++ cli/src/main/java/com/diffplug/spotless/cli/steps/generic/LicenseHeader.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.diffplug.spotless.cli.subcommands.steps.generic;
+package com.diffplug.spotless.cli.steps.generic;
 
 import java.io.File;
 import java.nio.file.Files;
